rename `gameplay` to `main` in graphics
