+++ package.json
@@ -109,7 +109,7 @@
         "height": 1080
       },
       {
-        "file": "gameplay.html",
+        "file": "main.html",
         "width": 1920,
         "height": 1080
       }
