+++ src/graphics/main.ts
@@ -1,6 +1,6 @@
 import { createHead } from '@vueuse/head';
 import { createApp } from 'vue';
-import App from './gameplay/main.vue';
+import App from './main/main.vue';
 // Import dev mode styling (will only activate if OSUFR_DEV_MODE is true)
 import '@4wc-stream-overlay/browser_shared/dev-mode';
